Removed SQL from BlockchainImpl
