+++ src/java/nxt/BlockchainImpl.java
@@ -1,15 +1,8 @@
 package nxt;
 
-import nxt.db.sql.Db;
-import nxt.db.sql.DbIterator;
-import nxt.db.sql.DbUtils;
 import nxt.db.BlockDb;
+import nxt.db.NxtIterator;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -73,110 +66,33 @@ final class BlockchainImpl implements Blockchain {
     }
 
     @Override
-    public DbIterator<BlockImpl> getAllBlocks() {
-        Connection con = null;
-        try {
-            con = Db.getConnection();
-            PreparedStatement pstmt = con.prepareStatement("SELECT * FROM block ORDER BY db_id ASC");
-            return getBlocks(con, pstmt);
-        } catch (SQLException e) {
-            DbUtils.close(con);
-            throw new RuntimeException(e.toString(), e);
-        }
+    public NxtIterator<BlockImpl> getAllBlocks() {
+        return Nxt.getStores().getBlockchainStore().getAllBlocks();
     }
     
     @Override
-    public DbIterator<BlockImpl> getBlocks(int from, int to) {
-        Connection con = null;
-        try {
-            con = Db.getConnection();
-            PreparedStatement pstmt = con.prepareStatement("SELECT * FROM block WHERE height <= ? AND height >= ? ORDER BY height DESC");
-            int blockchainHeight = getHeight();
-            pstmt.setInt(1, blockchainHeight - Math.max(from, 0));
-            pstmt.setInt(2, to > 0 ? blockchainHeight - to : 0);
-            return getBlocks(con, pstmt);
-        } catch (SQLException e) {
-            DbUtils.close(con);
-            throw new RuntimeException(e.toString(), e);
-        }
+    public NxtIterator<BlockImpl> getBlocks(int from, int to) {
+        return Nxt.getStores().getBlockchainStore().getBlocks(from, to);
     }
 
     @Override
-    public DbIterator<BlockImpl> getBlocks(Account account, int timestamp) {
+    public NxtIterator<BlockImpl> getBlocks(Account account, int timestamp) {
         return getBlocks(account, timestamp, 0, -1);
     }
 
     @Override
-    public DbIterator<BlockImpl> getBlocks(Account account, int timestamp, int from, int to) {
-        Connection con = null;
-        try {
-            con = Db.getConnection();
-            PreparedStatement pstmt = con.prepareStatement("SELECT * FROM block WHERE generator_id = ? "
-                    + (timestamp > 0 ? " AND timestamp >= ? " : " ") + "ORDER BY db_id DESC"
-                    + DbUtils.limitsClause(from, to));
-            int i = 0;
-            pstmt.setLong(++i, account.getId());
-            if (timestamp > 0) {
-                pstmt.setInt(++i, timestamp);
-            }
-            DbUtils.setLimits(++i, pstmt, from, to);
-            return getBlocks(con, pstmt);
-        } catch (SQLException e) {
-            DbUtils.close(con);
-            throw new RuntimeException(e.toString(), e);
-        }
-    }
-
-    @Override
-    public DbIterator<BlockImpl> getBlocks(Connection con, PreparedStatement pstmt) {
-        return new DbIterator<>(con, pstmt, new DbIterator.ResultSetReader<BlockImpl>() {
-            @Override
-            public BlockImpl get(Connection con, ResultSet rs) throws NxtException.ValidationException {
-                return blockDb.loadBlock(con, rs);
-            }
-        });
+    public NxtIterator<BlockImpl> getBlocks(Account account, int timestamp, int from, int to) {
+        return Nxt.getStores().getBlockchainStore().getBlocks(account, timestamp, from, to);
     }
 
     @Override
     public List<Long> getBlockIdsAfter(long blockId, int limit) {
-        if (limit > 1440) {
-            throw new IllegalArgumentException("Can't get more than 1440 blocks at a time");
-        }
-        try (Connection con = Db.getConnection();
-             PreparedStatement pstmt = con.prepareStatement("SELECT id FROM block WHERE db_id > (SELECT db_id FROM block WHERE id = ?) ORDER BY db_id ASC LIMIT ?")) {
-            List<Long> result = new ArrayList<>();
-            pstmt.setLong(1, blockId);
-            pstmt.setInt(2, limit);
-            try (ResultSet rs = pstmt.executeQuery()) {
-                while (rs.next()) {
-                    result.add(rs.getLong("id"));
-                }
-            }
-            return result;
-        } catch (SQLException e) {
-            throw new RuntimeException(e.toString(), e);
-        }
+        return Nxt.getStores().getBlockchainStore().getBlockIdsAfter(blockId, limit);
     }
 
     @Override
     public List<BlockImpl> getBlocksAfter(long blockId, int limit) {
-        if (limit > 1440) {
-            throw new IllegalArgumentException("Can't get more than 1440 blocks at a time");
-        }
-        try (Connection con = Db.getConnection();
-             PreparedStatement pstmt = con.prepareStatement("SELECT * FROM block WHERE db_id > (SELECT db_id FROM block WHERE id = ?) ORDER BY db_id ASC LIMIT ?")) {
-            List<BlockImpl> result = new ArrayList<>();
-            pstmt.setLong(1, blockId);
-            pstmt.setInt(2, limit);
-            try (ResultSet rs = pstmt.executeQuery()) {
-                while (rs.next()) {
-                    result.add(blockDb.loadBlock(con, rs));
-                }
-            }
-            return result;
-        } catch (NxtException.ValidationException|SQLException e) {
-            throw new RuntimeException(e.toString(), e);
-        }
+        return Nxt.getStores().getBlockchainStore().getBlocksAfter(blockId, limit);
     }
 
     @Override
@@ -225,119 +141,25 @@ final class BlockchainImpl implements Blockchain {
 
     @Override
     public int getTransactionCount() {
-        try (Connection con = Db.getConnection(); PreparedStatement pstmt = con.prepareStatement("SELECT COUNT(*) FROM transaction");
-             ResultSet rs = pstmt.executeQuery()) {
-            rs.next();
-            return rs.getInt(1);
-        } catch (SQLException e) {
-            throw new RuntimeException(e.toString(), e);
-        }
+        return Nxt.getStores().getBlockchainStore().getTransactionCount();
     }
 
     @Override
-    public DbIterator<TransactionImpl> getAllTransactions() {
-        Connection con = null;
-        try {
-            con = Db.getConnection();
-            PreparedStatement pstmt = con.prepareStatement("SELECT * FROM transaction ORDER BY db_id ASC");
-            return getTransactions(con, pstmt);
-        } catch (SQLException e) {
-            DbUtils.close(con);
-            throw new RuntimeException(e.toString(), e);
-        }
+    public NxtIterator<TransactionImpl> getAllTransactions() {
+        return Nxt.getStores().getBlockchainStore().getAllTransactions();
     }
 
     @Override
-    public DbIterator<TransactionImpl> getTransactions(Account account, byte type, byte subtype, int blockTimestamp) {
+    public NxtIterator<TransactionImpl> getTransactions(Account account, byte type, byte subtype, int blockTimestamp) {
         return getTransactions(account, 0, type, subtype, blockTimestamp, 0, -1);
     }
 
     @Override
-    public DbIterator<TransactionImpl> getTransactions(Account account, int numberOfConfirmations, byte type, byte subtype,
-                                                       int blockTimestamp, int from, int to) {
-        int height = numberOfConfirmations > 0 ? getHeight() - numberOfConfirmations : Integer.MAX_VALUE;
-        if (height < 0) {
-            throw new IllegalArgumentException("Number of confirmations required " + numberOfConfirmations
-                    + " exceeds current blockchain height " + getHeight());
-        }
-        Connection con = null;
-        try {
-            StringBuilder buf = new StringBuilder();
-            buf.append("SELECT * FROM transaction WHERE recipient_id = ? AND sender_id <> ? ");
-            if (blockTimestamp > 0) {
-                buf.append("AND block_timestamp >= ? ");
-            }
-            if (type >= 0) {
-                buf.append("AND type = ? ");
-                if (subtype >= 0) {
-                    buf.append("AND subtype = ? ");
-                }
-            }
-            if (height < Integer.MAX_VALUE) {
-                buf.append("AND height <= ? ");
-            }
-            buf.append("UNION ALL SELECT * FROM transaction WHERE sender_id = ? ");
-            if (blockTimestamp > 0) {
-                buf.append("AND block_timestamp >= ? ");
-            }
-            if (type >= 0) {
-                buf.append("AND type = ? ");
-                if (subtype >= 0) {
-                    buf.append("AND subtype = ? ");
-                }
-            }
-            if (height < Integer.MAX_VALUE) {
-                buf.append("AND height <= ? ");
-            }
-            buf.append("ORDER BY block_timestamp DESC, id DESC");
-            buf.append(DbUtils.limitsClause(from, to));
-            con = Db.getConnection();
-            PreparedStatement pstmt;
-            int i = 0;
-            pstmt = con.prepareStatement(buf.toString());
-            pstmt.setLong(++i, account.getId());
-            pstmt.setLong(++i, account.getId());
-            if (blockTimestamp > 0) {
-                pstmt.setInt(++i, blockTimestamp);
-            }
-            if (type >= 0) {
-                pstmt.setByte(++i, type);
-                if (subtype >= 0) {
-                    pstmt.setByte(++i, subtype);
-                }
-            }
-            if (height < Integer.MAX_VALUE) {
-                pstmt.setInt(++i, height);
-            }
-            pstmt.setLong(++i, account.getId());
-            if (blockTimestamp > 0) {
-                pstmt.setInt(++i, blockTimestamp);
-            }
-            if (type >= 0) {
-                pstmt.setByte(++i, type);
-                if (subtype >= 0) {
-                    pstmt.setByte(++i, subtype);
-                }
-            }
-            if (height < Integer.MAX_VALUE) {
-                pstmt.setInt(++i, height);
-            }
-            DbUtils.setLimits(++i, pstmt, from, to);
-            return getTransactions(con, pstmt);
-        } catch (SQLException e) {
-            DbUtils.close(con);
-            throw new RuntimeException(e.toString(), e);
-        }
+    public NxtIterator<TransactionImpl> getTransactions(Account account, int numberOfConfirmations, byte type, byte subtype,
+                                                        int blockTimestamp, int from, int to) {
+        return Nxt.getStores().getBlockchainStore().getTransactions(account, numberOfConfirmations, type, subtype, blockTimestamp, from, to);
+
     }
 
-    @Override
-    public DbIterator<TransactionImpl> getTransactions(Connection con, PreparedStatement pstmt) {
-        return new DbIterator<>(con, pstmt, new DbIterator.ResultSetReader<TransactionImpl>() {
-            @Override
-            public TransactionImpl get(Connection con, ResultSet rs) throws NxtException.ValidationException {
-                return transactionDb.loadTransaction(con, rs);
-            }
-        });
-    }
 
 }
