+++ src/java/nxt/db/mariadb/MariadbBlockchainStore.java
@@ -1,0 +1,6 @@
+package nxt.db.mariadb;
+
+import nxt.db.sql.SqlBlockchainStore;
+
+public class MariadbBlockchainStore extends SqlBlockchainStore {
+}
